api/requirements.txt
azure-functions
openai==1.3.5
tavily-python==0.3.3
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.4.2 